Adding the ability to pick the label applied to an automatic reference
This changes the syntax for `{{?RFCXXXX}}` and `{{!RFCBLAH}}` to allow
the author to specify a name for the reference.  This is added after a
space, as in `{{?RFC5389 STUN}}`.  This is especially handy for DOI
references, which I just added because it keeps the symbolic reference
tidy.  For example, `{{?DOI.10.1145/1282427.1282421 SST}}`, which is a
reference we are using in QUIC.

I added this ability throughout, which means that all references will
be affected.  I only have basically no testing on `<iref>` though.

This allowed me to consolidate the places where you were testing for
labels that started with a digit.  That too was necessary to have a
single location where the DOI reference was mangled.