Don't fetch with ?anchor=
I know that you commented this out because it was broken on
tools.ietf.org, but I have been meaning to propose this change for some
time.

Aside from causing bugs, including the anchor in the cache key destroys
cache efficiency.  Better off relying on the regex to update the anchor,
even if that is a tiny bit brittle.