1.3.37: contributors can be annotated with text
... that will be added to the Contributors section after their contact information.
The text is entered in the “contribution” entry of the contact
information for a contributor.
It is markdown and can contain paragraph breaks and other markup, as
shown below (with some author information as context, which is not
shown in the rendering).
Note that YAML has the “|” syntax that is strongly recommended for
entering markdown text into YAML entries.

author:
 - name: Carsten Bormann
   org: Universität Bremen TZI
   email: cabo@tzi.org
contributor:
 - name: Peter Occil
   email: poccil14 at gmail dot com
   contribution: |
     Peter Occil registered tags 30, 264, 265, 268–270
     ({{advanced-arithmetic}}), 38, 257, 266 and 267
     ({{domain-specific}}), and contributed much of the text about
     these tags in this document.
 - name: Jane Doe
   org: To do
   contribution: |
     Further contributors will be listed here as text is added.

     Plase stay tuned.