<xref format="counter"/>
This takes the bare section reference format of `{{1.3<RFC2119}}` and
makes it available for local use by dropping the external reference and
replacing the section number with a local section reference.

That is, where `{{foo}}` would produce `<xref target="foo"
format="default"/>`, `{{foo<}}` produces <xref target="foo"
format="counter"/>`.

Note: The previous code would generate the invalid `<xref target=""
section="foo" sectionFormat="bare"/>` for this case, so I believe that
this is safe.

Note: the pattern `[\w.]+` was used in adjacent code. I replaced this
with `.+`.