Expand BCP and STD references
The main problem that I see with this change is that it relies on tools.ietf.org.

If rfc-editor.org were to serve these, that would be much better, but
that doesn't currently happen.  I'm not tracking that work closely, so
maybe I've missed the feature request.

I "fixed" the RFC referencing part as well by adding leading zeros
automatically.  That was being done inconsistently throughout.

This only really works in v3 format documents as these references use
\<referencegroup>, so I did my best to disable that.