Use --spaces=1 for aasvg and protocol
Protocol is weird in that it produces a long line of "text" that aasvg
will put into the one element.  If you tell aasvg to break at the first
space (as opposed to the second), then it will place each letter more
precisely.  The cost is less control over inter-word spacing inside the
diagram (but better control over individual words).