Pass through "processing instructions" in v3
There are a small number of attributes on the `<rfc>` element that act
as processing instructions.  These determine if there is a TOC, that
sort of thing.

xml2rfc ignores other PIs when version 3 is enabled, so having these is
a useful way to tune its output.  Passing through attributes directly or
looking at the processing instructions is somewhat useful.

This suppresses processing instructions when producing v3 documents.