1.6.20: Ignore target attributes in RFC and I-D bibxml
Since version 1.3.33 [1], kramdown-rfc has had the ability to ignore target attributes in bibxml obtained for RFCs and I-Ds, allowing xml2rfc to fill in the proper values [1].
For the last couple of years, this needed to be triggered by setting the environment variable KRAMDOWN_NO_TARGETS.
This has been set as the default for Martin Thomson's i-d-template for a long time, and no adverse effects are known.
So with version 1.6.20, KRAMDOWN_NO_TARGETS is now the default; in a pinch use KRAMDOWN_KEEP_TARGETS to revert to previous behavior.

[1]: <https://mailarchive.ietf.org/arch/msg/rfc-markdown/8FKkKIMWSsUyy3s89wi3lMvgbUI> 2021-02-18