kramdown-rfc --help synopsis should indicate support for reading from stdin
kramdown-rfc uses [Ruby's
ARGF](https://docs.ruby-lang.org/en/master/ARGF.html), which accepts
data from stdin if no argument is supplied.

But the current synopsis seems to expect a filename as an argument.

The synopsis also appears to make some sort of arbitrary distinction
between a .md and a .mkd suffix, and adds a pipe metacharacter.

Simplify all of this by indicating that the filename is optional
(enclosing it in brackets) and renaming it from file.md|file.mkd to
MARKDOWNFILE.