1.6.41: Generate anchors for auto-generated references sections
Anchor-names:
  sec-normative-references
and
  sec-informative-references

(sec stands for section).
These currently are hard-wired; if this turns out to be a problem,
we'll add a configuration option to the front YAML.

There currently is no way to get an anchor for the v2v3-generated
umbrella section "References"; simply reference both subsections
instead if needed.