1.7.2: Handle section references into referencegroups
Up to now, it was really hard to reference into a referencegroup (STD
or BCP reference), as that reference could possibly contain multiple
RFCs, and one would need to be explicit which one was meant with the
section reference.

For RFC 9485, the RPC devised a way to express in RFCXML section
references into RFCs that are in reference groups.  This convention is
now supported by kramdown-rfc.

New syntax: RFC8949@STD94

This can be used for section references:

See {{Section 10 of RFC8996@BCP195}}.
See {{Section 10 ("Security Considerations") and 11 ("Privacy Considerations") of RFC8996@BCP195}}.

Nicknames (-xxx) and norm/inform flags only work to the right side of an @:

See {{Appendix D of RFC8949@!STD94}}.
See {{Appendix D of RFC8949@-cbor}}.

The left hand side of an @ MUST be an RFC number; breakage otherwise.
(Indirection would be great for the left hand side; for further study.)

For completeness, the @ syntax somewhat redundantly also works for
non-section references into reference groups:

See {{RFC8949@?STD94}}.
See {{RFC8949@STD94}}.