Add a way to suppress refcache fetching messages
These messages are not abnormal.  All other `stderr.puts` calls made in
this code relate to genuine problems that authors might need to pay
attention to (why those aren't using `warn` is something that might need
to be looked into) but the messages for downloading are not something
that anyone can do anything about.  This does not change the warning
that is emitted for failed downloads or other such errors.